backend/requirements.txt
# ----------------------------
# Core Django + REST Framework
# ----------------------------
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.4
drf-spectacular==0.26.5

# ----------------------------
# Authentication & Security
# ----------------------------
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1

# ----------------------------
# Database
# ----------------------------
psycopg2-binary==2.9.9
django-simple-history==3.4.0
django-import-export==3.3.1

# ----------------------------
# Environment Management
# ----------------------------
python-decouple==3.8

# ----------------------------
# Background Tasks / Async
# ----------------------------
celery==5.3.4
redis==5.0.1

# ----------------------------
# File Upload & Media
# ----------------------------
Pillow==10.1.0

# ----------------------------
# Utility & Development
# ----------------------------
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# ----------------------------
# Machine Learning & Data Science
# ----------------------------
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
statsmodels>=0.14.0
prophet>=1.1.4

# ----------------------------
# Optimization & Geospatial
# ----------------------------
networkx>=3.1
geopy>=2.3.0
requests>=2.31.0

# ----------------------------
# Data Validation & Processing
# ----------------------------
jsonschema>=4.18.0
python-dateutil==2.8.2
openpyxl==3.1.2

# ----------------------------
# OpenAI + LangChain
# ----------------------------
openai>=1.0.0
langchain>=0.2.0
langchain-community>=0.2.0
tiktoken>=0.7.0

# ----------------------------
# Production & Deployment
# ----------------------------
gunicorn>=21.2.0
whitenoise>=6.5.0
dj-database-url>=2.1.0
python-dotenv>=1.0.0

# ----------------------------
# Testing & Code Quality
# ----------------------------
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
flake8==6.1.0
black==23.11.0
